Correct the count of lines.
The number of lines in a file is usually computed at the number of newlines in the file minus 1, as described [here](https://codereview.stackexchange.com/questions/131143/counting-newlines-in-a-file).  I also confirmed this by running `wc` on linux with a file that is:

* empty
* non-empty but with no newlines
* with newlines and no trailing newline
* with newlines and with a trailing newline

I noticed this when I saw that the number of lines reported in the file list on coveralls didn't match the number of lines actually in the file.